vim-patch:8.1.2094: the fileio.c file is too big
Problem:    The fileio.c file is too big.
Solution:   Move buf_write() to bufwrite.c. (Yegappan Lakshmanan,
            closes vim/vim#4990)

https://github.com/vim/vim/commit/c079f0fed1c16495d726d616c5362edc04742a0d
