+++ src/nvim/bufwrite.h
@@ -1,0 +1,11 @@
+#ifndef NVIM_BUFWRITE_H
+#define NVIM_BUFWRITE_H
+
+#include "nvim/buffer_defs.h"
+#include "nvim/ex_cmds_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "bufwrite.h.generated.h"
+#endif
+
+#endif  // NVIM_BUFWRITE_H
